Move the TestServiceProvider into its own class (may be useful to other contributors)
